pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "GitHub repository monitoring tool with AI analysis"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Argus Contributors"}
]

dependencies = [
    "PyGithub==2.1.1",
    "python-dateutil==2.8.2",
    "pytz==2024.1",
    "requests==2.31.0",
    "PyYAML>=6.0",
]

[dependency-groups]
dev = [
    # 开发依赖可在此添加
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
